Make ANT_HOME configurable.
Changed the default log directory to be under /var/log.

Removed updating the installed.properties file and starting the anthill service
from the install script as this is done by Puppet.

Only install a single instance of the agent.

Ensure all configured directories exist before the install script is run.

Ensure that the bash package is installed.

Give an example with all the options.